+++ account_asset_management/__manifest__.py
@@ -4,10 +4,11 @@
 
 {
     "name": "Assets Management",
-    "version": "13.0.2.0.0",
+    "version": "13.0.3.7.3",
     "license": "AGPL-3",
     "depends": ["account", "report_xlsx_helper"],
     "excludes": ["account_asset"],
+    "development_status": "Production/Stable",
     "external_dependencies": {"python": ["python-dateutil"]},
     "author": "Noviat, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/account-financial-tools",
@@ -26,5 +27,6 @@
         "views/menuitem.xml",
         "data/cron.xml",
         "wizard/wiz_account_asset_report.xml",
+        "wizard/wiz_asset_move_reverse.xml",
     ],
 }