+++ account_move_line_purchase_info/__manifest__.py
@@ -5,12 +5,13 @@
 {
     "name": "Account Move Line Purchase Info",
     "summary": "Introduces the purchase order line to the journal items",
-    "version": "13.0.1.1.0",
+    "version": "13.0.1.1.3",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
-    "website": "https://www.github.com/OCA/account-financial-tools",
+    "website": "https://github.com/OCA/account-financial-tools",
     "category": "Generic",
     "depends": ["purchase_stock"],
     "license": "AGPL-3",
     "data": ["security/account_security.xml", "views/account_move_view.xml"],
     "installable": True,
+    "post_init_hook": "post_init_hook",
 }