+++ account_netting/__manifest__.py
@@ -4,12 +4,12 @@
 
 {
     "name": "Account netting",
-    "version": "13.0.1.0.0",
+    "version": "13.0.1.0.1",
     "summary": "Compensate AR/AP accounts from the same partner",
     "category": "Accounting & Finance",
     "author": "Tecnativa, " "Odoo Community Association (OCA)",
     "license": "AGPL-3",
-    "website": "https://github.com/OCA/account-financial-tools/",
+    "website": "https://github.com/OCA/account-financial-tools",
     "depends": ["account"],
     "data": ["wizards/account_move_make_netting_view.xml"],
     "installable": True,