+++ account_asset_management/readme/HISTORY.rst
@@ -1,3 +1,15 @@
+13.0.3.0.0 (2021-07-06)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* Allow to reverse the posting of a depreciation line instead of deleting the
+  journal entry.
+
+13.0.2.0.0 (2021-02-19)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* Add support for multi-company
+
+
 13.0.1.0.0 (2019-10-21)
 ~~~~~~~~~~~~~~~~~~~~~~~